+++ src/app/me/likes/Link.tsx
@@ -1,10 +1,10 @@
 import Link from "next/link";
 import React, { ComponentProps } from "react";
 
-export const YouMylistsPageLink: React.FC<
+export const YouLikesPageLink: React.FC<
   Omit<ComponentProps<typeof Link>, "href">
 > = ({ children, ...props }) => (
-  <Link href={"/you/mylists"} {...props}>
+  <Link href={"/me/likes"} {...props}>
     {children}
   </Link>
 );
