+++ src/app/me/likes/RegistrationsList.tsx
