+++ src/app/me/likes/layout.tsx
@@ -1,0 +1,3 @@
+import MylistsLayout from "../mylists/layout";
+
+export default MylistsLayout;
