+++ src/app/me/mylists/[id]/Details.tsx
