+++ src/app/me/mylists/[id]/RegistrationsList.tsx
