--- src/app/you/likes/layout.tsx
@@ -1,2 +1,0 @@
-import Layout from "../mylists/layout";
-export default Layout;
